requirements-dev.txt
-r requirements.txt
azure-functions~=1.19.0
mocker~=1.1.1
pytest~=8.2.1
pytest-cov~=5.0.0
pytest-mock~=3.14.0
requests-mock~=1.12.1
coverage~=7.5.1